requirements.txt
agno==2.1.1
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
ddgs==9.6.0
docstring_parser==0.17.0
fastapi==0.118.0
feedparser==6.0.12
filelock==3.19.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-genai==1.41.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
joblib==1.5.2
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
mdurl==0.1.2
newspaper4k==0.9.3.1
nltk==3.9.2
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
primp==0.15.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
rsa==4.9.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
starlette==0.48.0
tenacity==9.1.2
tldextract==5.3.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.23
uvicorn==0.37.0
websockets==15.0.1
